Attempted to consolidate a few state accesses across events.
Refactored the creditEvent, deliveredEvent, and timeoutEvent.

The deliveredEvent is now only invoked by ACKs.

The creditToBtxEvent now performs RMW ops to update the credit
state and the toBtx state. It also directly computes which
pkts to transmit and inserts the tx_msg_id along with the tx_pkts
bitmap into the scheduled_pkts_fifo. The scheduled_pkts_fifo
has replaced the active_msgs_fifo.

NOTE: if the scheduled_pkts_fifo is a PIFO this would enable some
programability of scheduling policies when transmitting pkts.
However, now that each entry contains multiple pkts, some flexibility
is lost, but I think this should be sufficient - especially if the
scheduling policy only depends on properties of the msg (as it does
for Homa).

The timeout event no longer needs to access the toBtx state, only
the delivered state and the max_tx_pkt_offset state.

The dequeue() operation in the Packetization module tracks the currently
active tx_msg_id and tx_pkts and doesn't need to touch any other
state.